Convert AndroidSceneRenderer to use SceneRenderer

+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/scene/FontMetricsService.java
@@ -3,11 +3,11 @@ package net.avh4.framework.uilayer.scene;
 import net.avh4.framework.uilayer.Font;
 
 public interface FontMetricsService {
-    int getAscent(Font font);
+    float getAscent(Font font);
 
-    int stringWidth(Font font, String text);
+    float stringWidth(Font font, String text);
 
-    int getDescent(Font font);
+    float getDescent(Font font);
 
-    int getLineHeight(Font font);
+    float getLineHeight(Font font);
 }
